refactor(core/component): [NotFound] move to app
