+++ packages/next-starter/src/app/components/AnimatedBG.ts
@@ -1,6 +1,6 @@
 import styled from 'styled-components';
 
-const AnimatedBG = styled.div`
+export const AnimatedBG = styled.div`
   position: absolute;
   width: 100vw;
   height: 100vh;
@@ -98,5 +98,3 @@ const AnimatedBG = styled.div`
     }
   }
 `;
-
-export default AnimatedBG;
